# public/openapi.yml
openapi: 3.0.0
info:
  title: INTERVEST.IO - API SPECS
  version: 0.0.1
  description: |
    Intervest.io adalah platform social investing sebagi tempat berkumpul trader & investor pasar modal sekaligus sebagi alat
    untuk membantu dalam menganalisa market.
servers:
  - url: 'http://localhost:10001'
    description: "localhost"
  - url: 'https://intervest.onbox.dev'
    description: "development"
...
paths:
...
  /auth/register:
    post:
      tags:
        - "Authentication"
      parameters:
        - $ref: "#/components/parameters/AppVersionParameter"
        - $ref: "#/components/parameters/AppPlatformParameter"
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequestBody'
      responses:
        200:
         description: ""
...
  /auth/login:
    post:
      parameters:
        - $ref: "#/components/parameters/AppVersionParameter"
        - $ref: "#/components/parameters/AppPlatformParameter"
        - $ref: "#/components/parameters/AppLangParameter"
      tags:
        - "Authentication"
      description: |
        `identity` bisa di isi dengan username, email atau phone.
      requestBody:
        content:
          application/json:
            example:
              identity: john@doe.com
              password: password
              device:
                id: '123123'
                name: 'Xiaomi Note 11 Pro 5G'
                brand: 'Xiaomi'
                os: Android 12
              coordinate: 
                longitude: -6.4013011623988305
                latitude: 106.85269506615691
            schema:
              $ref: '#/components/schemas/LoginRequestBody'
      responses:
        201:
         description: ""
         content:
           application/json:
            schema:
              $ref: '#/components/schemas/AuthSuccessResponse'
...
security:
  - ApiKey: []
  - AuthToken: []
components:
...
  parameters:
    AppVersionParameter:
      name: 'X-APP-VERSION'
      in: header
      required: true
      example: "1.0.0"
      schema:
        type: string
    AppPlatformParameter:
      name: 'X-APP-PLATFORM'
      in: header
      required: true
      example: "Android"
      schema:
        type: string
    AppLangParameter:
      name: 'X-APP-LANG'
      in: header
      required: true
      example: "en"
      schema:
        type: string
  securitySchemes:
    ApiKey:
      type: apiKey
      in: header
      description: "dev: `intervest-sandbox`"
      name: "x-api-key"
    AuthToken:
      type: http
      scheme: bearer
      bearerFormat: JWT    # optional, arbitrary value for documentation purposes
  schemas:
    LoginRequestBody:
      type: object
      properties:
        identity:
          type: string
          example: john@doe.com
          description: You can put email/username/phone
        password:
          type: string
        # provider:
        #   type: object
        #   description: |
        #     Paramter ini hanya dikirim padasaat melakukan autentikasi dengan 
        #     provider pihak ke 3 seperti google, facebook, twitter
        #   properties:
        #     name:
        #       type: string
        #       enum:
        #         - ["google", "facebook"]
        #     id:
        #       type: string
        #     token:
        #       type: string
        coordinate:
          type: object
          properties:
            latitude:
              type: number
            longitude:
              type: number
        device:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            brand:
              type: string
            os:
              type: string
    AuthSuccessResponse:
      type: object
      properties:
        status:
          type: string
        message:
          type: string
        data:
          type: object
          properties:
            tokens: 
              type: object
              description: "Authentication token (JWT) untuk di gunakan di endpoint terautentikasi."
              properties:
                access_token:
                  type: string
                  format: jwt
                refresh_token:
                  type: string
                  format: jwt
            user:
              type: object
              properties:
                id:
                  type: integer
                  format: int64
                  example: 12
                uuid:
                  type: string
                  format: uuid
                username:
                  type: string
                name:
                  type: string
                avatar:
                  type: string
                  nullable: true
                phone:
                  type: string
                phoneVerifiedAt:
                  type: string
                  format: date-time
                  nullable: true
                email:
                  type: string
                  format: email
                emailVerifiedAt:
                  type: string
                  format: date-time
                  nullable: true
                password:
                  type: string
                telegramAccount:
                  type: string
                  nullable: true
                telegramChatId:
                  type: string
                  nullable: true
                telegramVerifiedAt:
                  type: string
                  format: date-time
                  nullable: true
                memberId:
                  type: integer
                  format: int64
                memberPackageId:
                  type: integer
                  format: int64
                memberSubscriptionId:
                  type: integer
                  format: int64
                memberStatus:
                  type: string
                  nullable: true
...
    RegisterRequestBody:
      type: object
      properties:
        username:
          type: string
          example: "man.in.the.street"
        name:
          type: string
          description: "Full name"
          example: "Joe Blow"
        email:
          type: string
          example: "new1@mail.com"
        password:
          type: string
          example: "secret"
        phone:
          type: string
          example: +62857147333
        refferal_code:
          type: string
          example: ""
        coordinate:
          type: object
          properties:
            latitude:
              type: number
            longitude:
              type: number
          example:
            longitude: -6.4013011623988305
            latitude: 106.85269506615691
        device:
          type: object
          properties:
            id:
              type: string
            name:
              type: string
            brand:
              type: string
            os:
              type: string
          example:
            id: '123123'
            name: 'Xiaomi Note 11 Pro 5G'
            brand: 'Xiaomi'
            os: Android 12